third_party/blink/web_tests/http/tests/devtools/elements/styles/up-down-numerics-and-colors-expected.txt

Tests that numeric and color values are incremented/decremented correctly.


Running: testInit

Running: testAlterColor
    color: #EE3;

Running: testAlterNumber
    opacity: 11.6;

Running: testAlterBigNumber
    transform: rotate(1000000000000000065537deg);
